Rename AzureAdB2cConsoleApp to AzureAdB2cUserManagementConsoleApp
