+++ DotNetSample.sln
@@ -83,8 +83,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HostedServiceWebApp", "src\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AzureAdB2cWebApp", "src\AzureAdB2cWebApp\AzureAdB2cWebApp.csproj", "{42428887-28A6-4B9D-91FB-5F90D8C0D8EC}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AzureAdB2cConsoleApp", "src\AzureAdB2cConsoleApp\AzureAdB2cConsoleApp.csproj", "{5CE7B4D4-6FBA-462E-8A1B-68F9A17037A0}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DigitalSignatureConsoleApp", "src\DigitalSignatureConsoleApp\DigitalSignatureConsoleApp.csproj", "{44F219D2-5F99-44A3-9481-58FBE8AA6388}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HybridCryptoConsoleApp", "src\HybridCryptoConsoleApp\HybridCryptoConsoleApp.csproj", "{23E4E46B-86BB-486E-8EBC-831451729938}"
@@ -249,10 +247,6 @@ Global
 		{42428887-28A6-4B9D-91FB-5F90D8C0D8EC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{42428887-28A6-4B9D-91FB-5F90D8C0D8EC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{42428887-28A6-4B9D-91FB-5F90D8C0D8EC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5CE7B4D4-6FBA-462E-8A1B-68F9A17037A0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5CE7B4D4-6FBA-462E-8A1B-68F9A17037A0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5CE7B4D4-6FBA-462E-8A1B-68F9A17037A0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5CE7B4D4-6FBA-462E-8A1B-68F9A17037A0}.Release|Any CPU.Build.0 = Release|Any CPU
 		{44F219D2-5F99-44A3-9481-58FBE8AA6388}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{44F219D2-5F99-44A3-9481-58FBE8AA6388}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{44F219D2-5F99-44A3-9481-58FBE8AA6388}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -326,7 +320,6 @@ Global
 		{BB47EDD7-02B8-4BEB-A27A-F2376CB76A51} = {798EF2B5-D82E-49ED-8EDE-BD3B20040D59}
 		{7ECB039D-DD35-459A-A7B0-5A3DE3464AB2} = {512A5E50-BE18-4761-BA83-E56F4EC43B9E}
 		{42428887-28A6-4B9D-91FB-5F90D8C0D8EC} = {512A5E50-BE18-4761-BA83-E56F4EC43B9E}
-		{5CE7B4D4-6FBA-462E-8A1B-68F9A17037A0} = {798EF2B5-D82E-49ED-8EDE-BD3B20040D59}
 		{44F219D2-5F99-44A3-9481-58FBE8AA6388} = {798EF2B5-D82E-49ED-8EDE-BD3B20040D59}
 		{23E4E46B-86BB-486E-8EBC-831451729938} = {798EF2B5-D82E-49ED-8EDE-BD3B20040D59}
 		{D26261F7-D82A-49B1-BFDD-9FEFA6CA2EB7} = {512A5E50-BE18-4761-BA83-E56F4EC43B9E}
