+++ src/AzureAdB2cUserManagementConsoleApp/CustomAttributeNames.cs
@@ -1,4 +1,4 @@
-namespace AzureAdB2cConsoleApp;
+namespace AzureAdB2cUserManagementConsoleApp;
 
 public static class CustomAttributeNames {
 	public const string TestNumber = nameof(TestNumber);
