+++ src/AzureAdB2cUserManagementConsoleApp/GraphGetUserSample.cs
@@ -2,7 +2,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Microsoft.Graph;
 
-namespace AzureAdB2cConsoleApp;
+namespace AzureAdB2cUserManagementConsoleApp;
 
 // ユーザーをカスタム属性付きで取得
 public class GraphGetUserSample : GraphSampleBase {
