+++ src/AzureAdB2cUserManagementConsoleApp/GraphUpdateUserAccountEnabledSample.cs
@@ -2,25 +2,21 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Microsoft.Graph;
 
-namespace AzureAdB2cConsoleApp;
+namespace AzureAdB2cUserManagementConsoleApp;
 
 // ユーザーのカスタム属性を更新
-public class GraphUpdateUserCustomAttributeSample : GraphSampleBase {
-	public GraphUpdateUserCustomAttributeSample(IConfiguration config, ILogger<GraphSampleBase> logger) : base(config, logger) {
+public class GraphUpdateUserAccountEnabledSample : GraphSampleBase {
+	public GraphUpdateUserAccountEnabledSample(IConfiguration config, ILogger<GraphSampleBase> logger) : base(config, logger) {
 	}
 
 	protected override async Task RunCoreAsync(GraphServiceClient client) {
 		// todo:
 		var id = "{id}";
 
-		var attributeName = GetCustomAttributeFullName(CustomAttributeNames.TestNumber);
-		var attributeValue = new Random().Next(1, 10);
-
 		// 更新
 		var userToUpdate = new User {
-			AdditionalData = new Dictionary<string, object> {
-				[attributeName] = attributeValue,
-			},
+			// アカウントを無効にする
+			AccountEnabled = false,
 		};
 		await client.Users[id]
 			.Request()
@@ -29,7 +25,7 @@ public class GraphUpdateUserCustomAttributeSample : GraphSampleBase {
 		// 取得して確認
 		var userUpdated = await client.Users[id]
 			.Request()
-			.Select(attributeName)
+			.Select("accountEnabled")
 			.GetAsync();
 		ShowUser(userUpdated);
 	}
