+++ src/AzureAdB2cUserManagementConsoleApp/GraphUpdateUserCustomAttributeSample.cs
@@ -2,21 +2,25 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Microsoft.Graph;
 
-namespace AzureAdB2cConsoleApp;
+namespace AzureAdB2cUserManagementConsoleApp;
 
 // ユーザーのカスタム属性を更新
-public class GraphUpdateUserAccountEnabledSample : GraphSampleBase {
-	public GraphUpdateUserAccountEnabledSample(IConfiguration config, ILogger<GraphSampleBase> logger) : base(config, logger) {
+public class GraphUpdateUserCustomAttributeSample : GraphSampleBase {
+	public GraphUpdateUserCustomAttributeSample(IConfiguration config, ILogger<GraphSampleBase> logger) : base(config, logger) {
 	}
 
 	protected override async Task RunCoreAsync(GraphServiceClient client) {
 		// todo:
 		var id = "{id}";
 
+		var attributeName = GetCustomAttributeFullName(CustomAttributeNames.TestNumber);
+		var attributeValue = new Random().Next(1, 10);
+
 		// 更新
 		var userToUpdate = new User {
-			// アカウントを無効にする
-			AccountEnabled = false,
+			AdditionalData = new Dictionary<string, object> {
+				[attributeName] = attributeValue,
+			},
 		};
 		await client.Users[id]
 			.Request()
@@ -25,7 +29,7 @@ public class GraphUpdateUserAccountEnabledSample : GraphSampleBase {
 		// 取得して確認
 		var userUpdated = await client.Users[id]
 			.Request()
-			.Select("accountEnabled")
+			.Select(attributeName)
 			.GetAsync();
 		ShowUser(userUpdated);
 	}
