+++ src/AzureAdB2cUserManagementConsoleApp/GraphUpdateUserForceChangePasswordSample.cs
@@ -2,7 +2,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Microsoft.Graph;
 
-namespace AzureAdB2cConsoleApp;
+namespace AzureAdB2cUserManagementConsoleApp;
 
 // ユーザーを仮パスワードでリセット
 // 次回サインイン時に変更が必要
