+++ src/AzureAdB2cUserManagementConsoleApp/SampleService.cs
@@ -1,4 +1,4 @@
-
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
@@ -23,9 +23,15 @@ public class SampleService : IHostedService {
 
 		_lifetime.ApplicationStarted.Register(async () => {
 			try {
-				// todo:
-				Console.WriteLine("Run smaple");
-				await Task.CompletedTask;
+				await _services.GetRequiredService<GraphGetUserListSample>().RunAsync();
+				// ユーザー作成
+				//await _services.GetRequiredService<GraphCreateUserSample>().RunAsync();
+				// ユーザー更新（アカウントの有効・無効）
+				//await _services.GetRequiredService<GraphUpdateUserAccountEnabledSample>().RunAsync();
+				// ユーザー更新（カスタム属性）
+				//await _services.GetRequiredService<GraphUpdateUserCustomAttributeSample>().RunAsync();
+				// ユーザー更新（パスワードリセット）
+				//await _services.GetRequiredService<GraphUpdateUserForceChangePasswordSample>().RunAsync();
 			} catch (Exception) {
 				throw;
 			}
